requirements.txt
lxml
pandas
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pyttsx3
openai